all tests passing

+++ src/controllers/Posts.js
@@ -3,6 +3,8 @@ const { UserModel, PostModel } = require("../models");
 const createPost = async (req, res, next) => {
 	try {
 		const { id, title, description } = req.body;
+		if (!id || !title || !description)
+			return res.status(400).json({ msg: "Missing information" });
 		const post = new PostModel({
 			title,
 			description,
@@ -33,6 +35,8 @@ const readPost = async (req, res, next) => {
 const updatePost = async (req, res, next) => {
 	try {
 		const { id, title, description } = req.body;
+		if (!id || !title || !description)
+			return res.status(400).json({ msg: "Missing information" });
 		const originalUser = await PostModel.findById({ _id: id });
 		const user = await PostModel.updateOne(
 			{
@@ -56,6 +60,8 @@ const updatePost = async (req, res, next) => {
 const deletePost = async (req, res, next) => {
 	try {
 		const { id } = req.body;
+		if (!id || id.length < 10)
+			return res.status(400).json({ msg: "Please enter a valid id" });
 		const post = await PostModel.deleteOne({ _id: id });
 		res.json({ msg: "Post successfully deleted" });
 	} catch (error) {
